+++ core/src/com/unciv/ui/civilopedia/CivilopediaScreen.kt
@@ -7,9 +7,11 @@ import com.badlogic.gdx.scenes.scene2d.ui.*
 import com.unciv.Constants
 import com.unciv.UncivGame
 import com.unciv.models.ruleset.Ruleset
+import com.unciv.models.ruleset.VictoryType
 import com.unciv.models.stats.INamed
 import com.unciv.models.translations.tr
 import com.unciv.ui.utils.*
+import java.text.Collator
 import com.unciv.ui.utils.AutoScrollPane as ScrollPane
 
 /** Screen displaying the Civilopedia
@@ -96,7 +98,9 @@ class CivilopediaScreen(
         if (category !in categoryToEntries) return        // defense, allowing buggy panes to remain empty while others work
         var entries = categoryToEntries[category]!!
         if (category != CivilopediaCategories.Difficulty) // this is the only case where we need them in order
-            entries = entries.sortedBy { it.name.tr() }   // Alphabetical order of localized names
+            // Alphabetical order of localized names, using system default locale
+            entries = entries.sortedWith(compareBy(Collator.getInstance(), { it.name.tr() }))
+
         var currentY = -1f
 
         for (entry in entries) {
@@ -163,22 +167,31 @@ class CivilopediaScreen(
         val imageSize = 50f
         onBackButtonClicked { UncivGame.Current.setWorldScreen() }
 
+        val hideReligionItems = !game.gameInfo.hasReligionEnabled()
+        val noCulturalVictory = VictoryType.Cultural !in game.gameInfo.gameParameters.victoryTypes
+
         categoryToEntries[CivilopediaCategories.Building] = ruleset.buildings.values
-                .filter { "Will not be displayed in Civilopedia" !in it.uniques && !(it.isWonder || it.isNationalWonder) }
+                .filter { "Will not be displayed in Civilopedia" !in it.uniques
+                        && !(hideReligionItems && "Hidden when religion is disabled" in it.uniques)
+                        && !(noCulturalVictory && "Hidden when cultural victory is disabled" in it.uniques)
+                        && !it.isAnyWonder() }
                 .map {
                     CivilopediaEntry(
                         it.name,
-                        it.getDescription(false, null, ruleset),
+                        "",
                         CivilopediaCategories.Building.getImage?.invoke(it.name, imageSize),
                         (it as? ICivilopediaText).takeUnless { ct -> ct==null || ct.isCivilopediaTextEmpty() }
                     )
                 }
         categoryToEntries[CivilopediaCategories.Wonder] = ruleset.buildings.values
-                .filter { "Will not be displayed in Civilopedia" !in it.uniques && (it.isWonder || it.isNationalWonder) }
+                .filter { "Will not be displayed in Civilopedia" !in it.uniques
+                        && !(hideReligionItems && "Hidden when religion is disabled" in it.uniques)
+                        && !(noCulturalVictory && "Hidden when cultural victory is disabled" in it.uniques)
+                        && it.isAnyWonder() }
                 .map {
                     CivilopediaEntry(
                         it.name,
-                        it.getDescription(false, null, ruleset),
+                        "",
                         CivilopediaCategories.Wonder.getImage?.invoke(it.name, imageSize),
                         (it as? ICivilopediaText).takeUnless { ct -> ct==null || ct.isCivilopediaTextEmpty() }
                     )